src/imageassistantservice/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.117.1
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.23.4
google-cloud-aiplatform==1.38.1
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-logging==3.8.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.5
google-crc32c==1.7.1
google-genai==1.38.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.59.3
grpcio-health-checking==1.59.3
grpcio-status==1.59.3
grpcio-tools==1.59.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
numpy==2.3.3
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-grpc==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
